feat: move descriptors processing out of RX path
Processing of descriptors was taking a big chunk
total time spend in the RX packet processing.
Because if this we try to move big part of this
to preprocessing step performed when guest notifies
Firecracker about new descriptors available for RX queue.
This way when we process new packet, we can use already
preprocessed descriptor chains.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>