feat(queue): add pointers to access queue internals
Add pointers that allow us to access `DescriptorTable`,
`AvailRing` and `UsedRing` of the queue in the guest
memory without additional checks and accesses to
whole `GuestMemoryMmap`.
These pointers need to be set up during device activation
and all methods that access these pointers should only be
used after pointers are set. Because of this new requirement
a lot of test code needed to be updated to set these pointers
for the test queues.
Additionally when pointers are set up we mark memory as dirty.
We need to do this manually, because `GuestMemoryMmap` will
not be present during following memory accesses.
We also need to mark queue memory dirty after taking diff
snapshots as we clear the bitmaps after saving the guest
memory.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>