refactor(queue): update `is_valid` check for queue size
Previously this check was using number of available descriptors
in the queue to determine if the queue is valid. The problem here
is that this number changes at run time, so the guest can always
make the queue `len` bigger than the `max_size`. Instead we better
check the size guest told us about for verification.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>