feat: update the way test artifacts are downloaded and used
The old setup used `build/img` as the base directory
for artifacts pulled from S3. In S3 the artifacts directory
contains 1 subdirectory per architecture (x86_64 and aarch64).
This was causing final local artifacts path to be `build/img/x86_64`
or `build/img/aarch64`.

This commit changes the structure a to have a separate `build/artifacts`
directory with subdirectories containing different versions of
artifacts. The path to currently used artifacts will be placed in
`build/current_artifacts` file. This make is easy to switch between
multiple versions of artifacts without a need to delete/download them.
This also opens a door for A/B testing of artifacts.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>